Added null protection.

+++ orcid-integration-test/src/test/java/org/orcid/integration/blackbox/api/MemberV2Test.java
@@ -575,19 +575,19 @@ public class MemberV2Test {
         assertNotNull(activitiesResponse);
         ActivitiesSummary summary = activitiesResponse.getEntity(ActivitiesSummary.class);
         assertNotNull(summary);
-        if (!summary.getEducations().getSummaries().isEmpty()) {
+        if (summary.getEducations() != null && !summary.getEducations().getSummaries().isEmpty()) {
             for (EducationSummary education : summary.getEducations().getSummaries()) {
                 memberV2ApiClient.deleteEducationXml(user1OrcidId, education.getPutCode(), token);
             }
         }
 
-        if (!summary.getEmployments().getSummaries().isEmpty()) {
+        if (summary.getEmployments() != null && !summary.getEmployments().getSummaries().isEmpty()) {
             for (EmploymentSummary employment : summary.getEmployments().getSummaries()) {
                 memberV2ApiClient.deleteEmploymentXml(user1OrcidId, employment.getPutCode(), token);
             }
         }
 
-        if (!summary.getFundings().getFundingGroup().isEmpty()) {
+        if (summary.getFundings() != null && !summary.getFundings().getFundingGroup().isEmpty()) {
             for (FundingGroup group : summary.getFundings().getFundingGroup()) {
                 for (FundingSummary funding : group.getFundingSummary()) {
                     memberV2ApiClient.deleteFundingXml(user1OrcidId, funding.getPutCode(), token);
@@ -595,7 +595,7 @@ public class MemberV2Test {
             }
         }
 
-        if (!summary.getWorks().getWorkGroup().isEmpty()) {
+        if (summary.getWorks() != null && !summary.getWorks().getWorkGroup().isEmpty()) {
             for (WorkGroup group : summary.getWorks().getWorkGroup()) {
                 for (WorkSummary work : group.getWorkSummary()) {
                     memberV2ApiClient.deleteWorkXml(user1OrcidId, work.getPutCode(), token);
